Auth_Backend/requirements.txt
fastapi
uvicorn
python-jose
passlib[bcrypt]
sqlalchemy
pydantic
python-dotenv